+++ src/app/matx-autocomplete-demo/matx-autocomplete-demo.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { of } from 'rxjs';
 import { delay } from 'rxjs/operators';
-import { FormBuilder, FormGroup } from '@angular/forms';
+import { FormBuilder, FormGroup, NgForm, Validators } from '@angular/forms';
 
 @Component({
   selector: 'app-autocomplete-demo',
@@ -22,11 +22,12 @@ export class MatxAutocompleteDemoComponent implements OnInit {
   ];
 
   form1Model = {
-    autocompleteInputDisabled: {name: 'Alabama'},
     autocompleteInput0: {name: 'Alabama'},
     autocompleteInput1: 'Florida',
     autocompleteInput2: {name: 'Alabama'},
-    autocompleteInput3: [{name: 'Florida'}]
+    autocompleteInput3: [{name: 'Florida'}],
+    autocompleteInput4: ['Alabama'],
+    autocompleteInput5: [{name: 'Alabama'}]
   };
 
   form2: FormGroup;
@@ -36,7 +37,9 @@ export class MatxAutocompleteDemoComponent implements OnInit {
   ngOnInit(): void {
     this.form2 = this.fb.group({
       autocompleteInput1: 'Alabama',
-      autocompleteInput2: {name: 'Florida'}
+      autocompleteInput2: {name: 'Florida'},
+      autocompleteInput3: [['Florida'], Validators.minLength(2)],
+      autocompleteInput4: [[{name: 'Alabama'}], Validators.minLength(2)]
     });
   }
 
@@ -50,5 +53,15 @@ export class MatxAutocompleteDemoComponent implements OnInit {
     // console.log('filterBy02 - value: ', value);
     return of([{name: 'Alabama'}, {name: 'Florida'}].filter(it => it.name.toLowerCase().includes(value.toLowerCase())))
       .pipe(delay(4000));
+  }
+
+  checkForm1(form: NgForm) {
+    console.log('checking...');
+    console.log('form.invalid: ', form.invalid);
+  }
+
+  checkForm2() {
+    console.log('checking...');
+    console.log('this.form2.invalid: ', this.form2.invalid);
   }
 }